refactor(accounts): split into multiple files
Split the monolithic accounts.rs (6480 lines) into a module directory
with three focused files:

- mod.rs: Type definitions, structs, enums, and Account impl (502 lines)
- login.rs: Login/logout flows and account management methods (5127 lines)
- setup.rs: Account setup infrastructure and relay configuration (893 lines)

Methods called cross-file within the module use pub(super) visibility.
No logic changes — pure mechanical reorganization.
