+++ src/whitenoise/accounts/login.rs
@@ -1,486 +1,14 @@
-use std::collections::HashSet;
-use std::fmt;
-use std::path::Path;
-use std::str::FromStr;
-
-use chrono::{DateTime, Utc};
-use mdk_core::prelude::*;
-use mdk_sqlite_storage::MdkSqliteStorage;
-use nostr_blossom::client::BlossomClient;
 use nostr_sdk::prelude::*;
-use serde::{Deserialize, Serialize};
-use thiserror::Error;
 
+use super::{
+    Account, AccountType, DiscoveredRelayLists, ExternalSignerRelaySetup, LoginError, LoginResult,
+    LoginStatus,
+};
 use crate::RelayType;
-use crate::nostr_manager::{NostrManager, NostrManagerError};
-use crate::types::ImageType;
-use crate::whitenoise::database::published_key_packages::PublishedKeyPackage;
 use crate::whitenoise::error::Result;
 use crate::whitenoise::relays::Relay;
 use crate::whitenoise::users::User;
 use crate::whitenoise::{Whitenoise, WhitenoiseError};
-
-/// The type of account authentication.
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Default, Serialize, Deserialize)]
-pub enum AccountType {
-    /// Account with locally stored private key.
-    #[default]
-    Local,
-    /// Account using external signer (e.g., Amber via NIP-55).
-    /// The private key never touches this app.
-    External,
-}
-
-impl fmt::Display for AccountType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            AccountType::Local => write!(f, "local"),
-            AccountType::External => write!(f, "external"),
-        }
-    }
-}
-
-impl FromStr for AccountType {
-    type Err = String;
-
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "local" => Ok(AccountType::Local),
-            "external" => Ok(AccountType::External),
-            _ => Err(format!("Unknown account type: {}", s)),
-        }
-    }
-}
-
-/// The status of a login attempt.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum LoginStatus {
-    /// Login completed successfully. Account is fully activated with relay lists,
-    /// subscriptions, and a published key package.
-    Complete,
-    /// Relay lists were not found on the network. The account exists in a partial
-    /// state and the caller must resolve relay lists before login can complete.
-    /// Use `login_publish_default_relays` or `login_with_custom_relay` to continue,
-    /// or `login_cancel` to clean up.
-    NeedsRelayLists,
-}
-
-/// The result of a login attempt.
-#[derive(Debug, Clone)]
-pub struct LoginResult {
-    /// The account that was created or found.
-    pub account: Account,
-    /// Whether login completed or needs further action.
-    pub status: LoginStatus,
-}
-
-/// The three relay lists discovered during login.
-///
-/// Each field holds the relays found on the network for that list type.
-/// An empty `Vec` means that list was **not found** and must be published
-/// (either from defaults or from a user-provided relay) before login can
-/// complete.  Use [`DiscoveredRelayLists::is_complete`] to check whether all
-/// three are present.
-#[derive(Debug, Clone)]
-pub struct DiscoveredRelayLists {
-    /// NIP-65 relay list (kind 10002).  Empty if not found on the network.
-    pub nip65: Vec<Relay>,
-    /// Inbox relays (kind 10050).  Empty if not found on the network.
-    pub inbox: Vec<Relay>,
-    /// Key-package relays (kind 10051).  Empty if not found on the network.
-    pub key_package: Vec<Relay>,
-}
-
-impl DiscoveredRelayLists {
-    /// Returns `true` when all three relay lists were found on the network.
-    pub fn is_complete(&self) -> bool {
-        !self.nip65.is_empty() && !self.inbox.is_empty() && !self.key_package.is_empty()
-    }
-
-    /// Returns the relay slice for the given `relay_type`.
-    pub fn relays(&self, relay_type: RelayType) -> &[Relay] {
-        match relay_type {
-            RelayType::Nip65 => &self.nip65,
-            RelayType::Inbox => &self.inbox,
-            RelayType::KeyPackage => &self.key_package,
-        }
-    }
-
-    /// Returns the relays for `relay_type` when non-empty, otherwise `fallback`.
-    pub fn relays_or<'a>(&'a self, relay_type: RelayType, fallback: &'a [Relay]) -> &'a [Relay] {
-        let found = self.relays(relay_type);
-        if found.is_empty() { fallback } else { found }
-    }
-
-    /// Merge `other` into `self`, keeping any non-empty field from either side.
-    ///
-    /// Each field is updated only when the current value is empty and the
-    /// incoming value is non-empty, so previously discovered relays are never
-    /// discarded.
-    pub fn merge(&mut self, other: Self) {
-        if self.nip65.is_empty() && !other.nip65.is_empty() {
-            self.nip65 = other.nip65;
-        }
-        if self.inbox.is_empty() && !other.inbox.is_empty() {
-            self.inbox = other.inbox;
-        }
-        if self.key_package.is_empty() && !other.key_package.is_empty() {
-            self.key_package = other.key_package;
-        }
-    }
-}
-
-/// Errors specific to the login flow.
-#[derive(Debug, Error)]
-pub enum LoginError {
-    /// The provided private key is not valid (bad format, bad encoding, etc.).
-    #[error("Invalid private key format: {0}")]
-    InvalidKeyFormat(String),
-
-    /// Could not connect to any relay to fetch or publish data.
-    #[error("Failed to connect to any relays")]
-    NoRelayConnections,
-
-    /// The operation timed out.
-    #[error("Login operation timed out: {0}")]
-    Timeout(String),
-
-    /// No partial login in progress for the given pubkey.
-    #[error("No login in progress for this account")]
-    NoLoginInProgress,
-
-    /// An internal error that doesn't fit the above categories.
-    #[error("Login error: {0}")]
-    Internal(String),
-}
-
-impl From<nostr_sdk::key::Error> for LoginError {
-    fn from(err: nostr_sdk::key::Error) -> Self {
-        Self::InvalidKeyFormat(err.to_string())
-    }
-}
-
-impl From<WhitenoiseError> for LoginError {
-    fn from(err: WhitenoiseError) -> Self {
-        match err {
-            WhitenoiseError::NostrManager(NostrManagerError::NoRelayConnections) => {
-                Self::NoRelayConnections
-            }
-            WhitenoiseError::NostrManager(NostrManagerError::Timeout) => {
-                Self::Timeout("relay operation timed out".to_string())
-            }
-            other => Self::Internal(other.to_string()),
-        }
-    }
-}
-
-#[derive(Debug, Error)]
-pub enum AccountError {
-    #[error("Failed to parse public key: {0}")]
-    PublicKeyError(#[from] nostr_sdk::key::Error),
-
-    #[error("Failed to initialize Nostr manager: {0}")]
-    NostrManagerError(#[from] NostrManagerError),
-
-    #[error("Nostr MLS error: {0}")]
-    NostrMlsError(#[from] mdk_core::Error),
-
-    #[error("Nostr MLS SQLite storage error: {0}")]
-    NostrMlsSqliteStorageError(#[from] mdk_sqlite_storage::error::Error),
-
-    #[error("Nostr MLS not initialized")]
-    NostrMlsNotInitialized,
-
-    #[error("Whitenoise not initialized")]
-    WhitenoiseNotInitialized,
-}
-
-/// Result of setting up relays for an external signer account.
-///
-/// Contains the relay lists for each type and flags indicating which lists
-/// need to be published (when defaults were used because no existing relays
-/// were found on the network).
-#[derive(Debug)]
-pub(crate) struct ExternalSignerRelaySetup {
-    pub nip65_relays: Vec<Relay>,
-    pub inbox_relays: Vec<Relay>,
-    pub key_package_relays: Vec<Relay>,
-    pub should_publish_nip65: bool,
-    pub should_publish_inbox: bool,
-    pub should_publish_key_package: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct Account {
-    pub id: Option<i64>,
-    pub pubkey: PublicKey,
-    pub user_id: i64,
-    /// The type of account (local key or external signer).
-    pub account_type: AccountType,
-    pub last_synced_at: Option<DateTime<Utc>>,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-}
-
-impl Account {
-    /// Returns true if this account uses an external signer.
-    pub fn uses_external_signer(&self) -> bool {
-        matches!(self.account_type, AccountType::External)
-    }
-
-    /// Returns true if this account has a locally stored private key.
-    pub fn has_local_key(&self) -> bool {
-        matches!(self.account_type, AccountType::Local)
-    }
-}
-
-impl Account {
-    pub(crate) async fn new(
-        whitenoise: &Whitenoise,
-        keys: Option<Keys>,
-    ) -> Result<(Account, Keys)> {
-        let keys = keys.unwrap_or_else(Keys::generate);
-
-        let (user, _created) =
-            User::find_or_create_by_pubkey(&keys.public_key(), &whitenoise.database).await?;
-
-        let account = Account {
-            id: None,
-            user_id: user.id.unwrap(),
-            pubkey: keys.public_key(),
-            account_type: AccountType::Local,
-            last_synced_at: None,
-            created_at: Utc::now(),
-            updated_at: Utc::now(),
-        };
-
-        Ok((account, keys))
-    }
-
-    /// Creates a new account for an external signer (pubkey only, no private key).
-    pub(crate) async fn new_external(
-        whitenoise: &Whitenoise,
-        pubkey: PublicKey,
-    ) -> Result<Account> {
-        let (user, _created) =
-            User::find_or_create_by_pubkey(&pubkey, &whitenoise.database).await?;
-
-        let account = Account {
-            id: None,
-            user_id: user.id.unwrap(),
-            pubkey,
-            account_type: AccountType::External,
-            last_synced_at: None,
-            created_at: Utc::now(),
-            updated_at: Utc::now(),
-        };
-
-        Ok(account)
-    }
-
-    /// Convert last_synced_at to a Timestamp applying a lookback buffer.
-    /// Clamps future timestamps to now to avoid empty subscriptions.
-    /// Returns None if the account has never synced.
-    pub(crate) fn since_timestamp(&self, buffer_secs: u64) -> Option<nostr_sdk::Timestamp> {
-        let ts = self.last_synced_at?;
-        // Clamp to now, then apply buffer
-        let now_secs = Utc::now().timestamp().max(0) as u64;
-        let last_secs = (ts.timestamp().max(0) as u64).min(now_secs);
-        let secs = last_secs.saturating_sub(buffer_secs);
-        Some(nostr_sdk::Timestamp::from(secs))
-    }
-
-    /// Retrieves the account's configured relays for a specific relay type.
-    ///
-    /// This method fetches the locally cached relays associated with this account
-    /// for the specified relay type. Different relay types serve different purposes
-    /// in the Nostr ecosystem and are published as separate relay list events.
-    ///
-    /// # Arguments
-    ///
-    /// * `relay_type` - The type of relays to retrieve:
-    ///   - `RelayType::Nip65` - General purpose relays for reading/writing events (kind 10002)
-    ///   - `RelayType::Inbox` - Specialized relays for receiving private messages (kind 10050)
-    ///   - `RelayType::KeyPackage` - Relays that store MLS key packages (kind 10051)
-    /// * `whitenoise` - The Whitenoise instance for database operations
-    pub async fn relays(
-        &self,
-        relay_type: RelayType,
-        whitenoise: &Whitenoise,
-    ) -> Result<Vec<Relay>> {
-        let user = self.user(&whitenoise.database).await?;
-        let relays = user.relays(relay_type, &whitenoise.database).await?;
-        Ok(relays)
-    }
-
-    /// Helper method to retrieve the NIP-65 relays for this account.
-    pub(crate) async fn nip65_relays(&self, whitenoise: &Whitenoise) -> Result<Vec<Relay>> {
-        let user = self.user(&whitenoise.database).await?;
-        let relays = user.relays(RelayType::Nip65, &whitenoise.database).await?;
-        Ok(relays)
-    }
-
-    /// Helper method to retrieve the inbox relays for this account.
-    pub(crate) async fn inbox_relays(&self, whitenoise: &Whitenoise) -> Result<Vec<Relay>> {
-        let user = self.user(&whitenoise.database).await?;
-        let relays = user.relays(RelayType::Inbox, &whitenoise.database).await?;
-        Ok(relays)
-    }
-
-    /// Helper method to retrieve the key package relays for this account.
-    pub(crate) async fn key_package_relays(&self, whitenoise: &Whitenoise) -> Result<Vec<Relay>> {
-        let user = self.user(&whitenoise.database).await?;
-        let relays = user
-            .relays(RelayType::KeyPackage, &whitenoise.database)
-            .await?;
-        Ok(relays)
-    }
-
-    /// Adds a relay to the account's relay list for the specified relay type.
-    ///
-    /// This method adds a relay to the account's local relay configuration and automatically
-    /// publishes the updated relay list to the Nostr network. The relay will be associated
-    /// with the specified type (NIP-65, Inbox, or Key Package relays) and become part of
-    /// the account's relay configuration for that purpose.
-    ///
-    /// # Arguments
-    ///
-    /// * `relay` - The relay to add to the account's relay list
-    /// * `relay_type` - The type of relay list to add this relay to:
-    ///   - `RelayType::Nip65` - General purpose relays (kind 10002)
-    ///   - `RelayType::Inbox` - Inbox relays for private messages (kind 10050)
-    ///   - `RelayType::KeyPackage` - Key package relays for MLS (kind 10051)
-    /// * `whitenoise` - The Whitenoise instance for database and network operations
-    pub async fn add_relay(
-        &self,
-        relay: &Relay,
-        relay_type: RelayType,
-        whitenoise: &Whitenoise,
-    ) -> Result<()> {
-        let user = self.user(&whitenoise.database).await?;
-        user.add_relay(relay, relay_type, &whitenoise.database)
-            .await?;
-        whitenoise
-            .nostr
-            .ensure_relays_connected(std::slice::from_ref(&relay.url))
-            .await?;
-        whitenoise
-            .background_publish_account_relay_list(self, relay_type, None)
-            .await?;
-        tracing::debug!(target: "whitenoise::accounts", "Added relay to account: {:?}", relay.url);
-
-        Ok(())
-    }
-
-    /// Removes a relay from the account's relay list for the specified relay type.
-    ///
-    /// This method removes a relay from the account's local relay configuration and automatically
-    /// publishes the updated relay list to the Nostr network. The relay will be disassociated
-    /// from the specified type and the account will stop using it for that purpose.
-    ///
-    /// # Arguments
-    ///
-    /// * `relay` - The relay to remove from the account's relay list
-    /// * `relay_type` - The type of relay list to remove this relay from:
-    ///   - `RelayType::Nip65` - General purpose relays (kind 10002)
-    ///   - `RelayType::Inbox` - Inbox relays for private messages (kind 10050)
-    ///   - `RelayType::KeyPackage` - Key package relays for MLS (kind 10051)
-    /// * `whitenoise` - The Whitenoise instance for database and network operations
-    pub async fn remove_relay(
-        &self,
-        relay: &Relay,
-        relay_type: RelayType,
-        whitenoise: &Whitenoise,
-    ) -> Result<()> {
-        let user = self.user(&whitenoise.database).await?;
-        user.remove_relay(relay, relay_type, &whitenoise.database)
-            .await?;
-        whitenoise
-            .background_publish_account_relay_list(self, relay_type, None)
-            .await?;
-        tracing::debug!(target: "whitenoise::accounts", "Removed relay from account: {:?}", relay.url);
-        Ok(())
-    }
-
-    /// Retrieves the cached metadata for this account.
-    ///
-    /// This method returns the account's stored metadata from the local database without
-    /// performing any network requests. The metadata contains profile information such as
-    /// display name, about text, picture URL, and other profile fields as defined by NIP-01.
-    ///
-    /// # Arguments
-    ///
-    /// * `whitenoise` - The Whitenoise instance used to access the database
-    pub async fn metadata(&self, whitenoise: &Whitenoise) -> Result<Metadata> {
-        let user = self.user(&whitenoise.database).await?;
-        Ok(user.metadata.clone())
-    }
-
-    /// Updates the account's metadata with new values and publishes to the network.
-    ///
-    /// This method updates the account's metadata in the local database with the provided
-    /// values and automatically publishes a metadata event (kind 0) to the account's relays.
-    /// This allows other users and clients to see the updated profile information.
-    ///
-    /// # Arguments
-    ///
-    /// * `metadata` - The new metadata to set for this account
-    /// * `whitenoise` - The Whitenoise instance for database and network operations
-    pub async fn update_metadata(
-        &self,
-        metadata: &Metadata,
-        whitenoise: &Whitenoise,
-    ) -> Result<()> {
-        tracing::debug!(target: "whitenoise::accounts", "Updating metadata for account: {:?}", self.pubkey);
-        let mut user = self.user(&whitenoise.database).await?;
-        user.metadata = metadata.clone();
-        user.save(&whitenoise.database).await?;
-        whitenoise.background_publish_account_metadata(self).await?;
-        Ok(())
-    }
-
-    /// Uploads an image file to a Blossom server and returns the URL.
-    ///
-    /// # Arguments
-    /// * `file_path` - Path to the image file to upload
-    /// * `image_type` - Image type (JPEG, PNG, etc.)
-    /// * `server` - Blossom server URL
-    /// * `whitenoise` - Whitenoise instance for accessing account keys
-    pub async fn upload_profile_picture(
-        &self,
-        file_path: &str,
-        image_type: ImageType,
-        server: Url,
-        whitenoise: &Whitenoise,
-    ) -> Result<String> {
-        let client = BlossomClient::new(server);
-        let signer = whitenoise.get_signer_for_account(self)?;
-        let data = tokio::fs::read(file_path).await?;
-
-        let descriptor = client
-            .upload_blob(
-                data,
-                Some(image_type.mime_type().to_string()),
-                None,
-                Some(&signer),
-            )
-            .await
-            .map_err(|err| WhitenoiseError::Other(anyhow::anyhow!(err)))?;
-
-        Ok(descriptor.url.to_string())
-    }
-
-    pub(crate) fn create_mdk(
-        pubkey: PublicKey,
-        data_dir: &Path,
-        keyring_service_id: &str,
-    ) -> core::result::Result<MDK<MdkSqliteStorage>, AccountError> {
-        let mls_storage_dir = data_dir.join("mls").join(pubkey.to_hex());
-        let db_key_id = format!("mdk.db.key.{}", pubkey.to_hex());
-        let storage = MdkSqliteStorage::new(mls_storage_dir, keyring_service_id, &db_key_id)?;
-        Ok(MDK::new(storage))
-    }
-}
 
 impl Whitenoise {
     /// Creates a new identity (account) for the user.
@@ -1637,743 +1165,6 @@ impl Whitenoise {
     pub async fn find_account_by_pubkey(&self, pubkey: &PublicKey) -> Result<Account> {
         Account::find_by_pubkey(pubkey, &self.database).await
     }
-
-    async fn create_base_account_with_private_key(&self, keys: &Keys) -> Result<Account> {
-        let (account, _keys) = Account::new(self, Some(keys.clone())).await?;
-
-        self.secrets_store.store_private_key(keys).map_err(|e| {
-            tracing::error!(target: "whitenoise::accounts", "Failed to store private key: {}", e);
-            e
-        })?;
-
-        let account = self.persist_account(&account).await?;
-
-        Ok(account)
-    }
-
-    async fn activate_account(
-        &self,
-        account: &Account,
-        user: &User,
-        is_new_account: bool,
-        nip65_relays: &[Relay],
-        inbox_relays: &[Relay],
-        key_package_relays: &[Relay],
-    ) -> Result<()> {
-        // Create a fresh cancellation channel for this account's background tasks.
-        // Any previous channel (from a prior login) is replaced, which also drops
-        // any stale receivers.
-        let (cancel_tx, _) = tokio::sync::watch::channel(false);
-        self.background_task_cancellation
-            .insert(account.pubkey, cancel_tx);
-
-        let relay_urls: Vec<RelayUrl> = Relay::urls(
-            nip65_relays
-                .iter()
-                .chain(inbox_relays)
-                .chain(key_package_relays),
-        );
-        self.nostr.ensure_relays_connected(&relay_urls).await?;
-        tracing::debug!(target: "whitenoise::accounts", "Relays connected");
-        if let Err(e) = self.refresh_global_subscription_for_user(user).await {
-            tracing::warn!(
-                target: "whitenoise::accounts",
-                "Failed to refresh global subscription for new user {}: {}",
-                user.pubkey,
-                e
-            );
-        }
-        tracing::debug!(target: "whitenoise::accounts", "Global subscription refreshed for account user");
-        self.setup_subscriptions(account, nip65_relays, inbox_relays)
-            .await?;
-        tracing::debug!(target: "whitenoise::accounts", "Subscriptions setup");
-        self.setup_key_package(account, is_new_account, key_package_relays)
-            .await?;
-        tracing::debug!(target: "whitenoise::accounts", "Key package setup");
-        Ok(())
-    }
-
-    /// Activates an account without publishing anything (for external signer accounts).
-    /// This sets up relay connections and subscriptions but skips key package publishing.
-    async fn activate_account_without_publishing(
-        &self,
-        account: &Account,
-        user: &User,
-        nip65_relays: &[Relay],
-        inbox_relays: &[Relay],
-        key_package_relays: &[Relay],
-    ) -> Result<()> {
-        let (cancel_tx, _) = tokio::sync::watch::channel(false);
-        self.background_task_cancellation
-            .insert(account.pubkey, cancel_tx);
-
-        let relay_urls: Vec<RelayUrl> = Relay::urls(
-            nip65_relays
-                .iter()
-                .chain(inbox_relays)
-                .chain(key_package_relays),
-        );
-        self.nostr.ensure_relays_connected(&relay_urls).await?;
-        tracing::debug!(target: "whitenoise::accounts", "Relays connected");
-
-        if let Err(e) = self.refresh_global_subscription_for_user(user).await {
-            tracing::warn!(
-                target: "whitenoise::accounts",
-                "Failed to refresh global subscription for new user {}: {}",
-                user.pubkey,
-                e
-            );
-        }
-        tracing::debug!(target: "whitenoise::accounts", "Global subscription refreshed");
-
-        self.setup_subscriptions(account, nip65_relays, inbox_relays)
-            .await?;
-        tracing::debug!(target: "whitenoise::accounts", "Subscriptions setup");
-
-        // Note: We skip key package setup for external signer accounts.
-        // Key packages need to be published separately with the external signer.
-        tracing::debug!(target: "whitenoise::accounts", "Skipping key package setup (external signer)");
-
-        Ok(())
-    }
-
-    async fn persist_account(&self, account: &Account) -> Result<Account> {
-        let saved_account = account.save(&self.database).await.map_err(|e| {
-            tracing::error!(target: "whitenoise::accounts", "Failed to save account: {}", e);
-            // Try to clean up stored private key
-            if let Err(cleanup_err) = self.secrets_store.remove_private_key_for_pubkey(&account.pubkey) {
-                tracing::error!(target: "whitenoise::accounts", "Failed to cleanup private key after account save failure: {}", cleanup_err);
-            }
-            e
-        })?;
-        tracing::debug!(target: "whitenoise::accounts", "Account saved to database");
-        Ok(saved_account)
-    }
-
-    async fn setup_key_package(
-        &self,
-        account: &Account,
-        is_new_account: bool,
-        key_package_relays: &[Relay],
-    ) -> Result<()> {
-        let mut needs_publish = true;
-
-        if !is_new_account {
-            tracing::debug!(target: "whitenoise::accounts", "Found {} key package relays", key_package_relays.len());
-            let relays_urls = Relay::urls(key_package_relays);
-
-            if let Some(event) = self
-                .nostr
-                .fetch_user_key_package(account.pubkey, &relays_urls)
-                .await?
-            {
-                if self.is_own_key_package(&account.pubkey, &event.id).await {
-                    tracing::debug!(
-                        target: "whitenoise::accounts",
-                        "Found existing key package with live key material, skipping publish"
-                    );
-                    needs_publish = false;
-                } else {
-                    tracing::debug!(
-                        target: "whitenoise::accounts",
-                        "Key package on relay was not published by this client or key material is gone, publishing new one"
-                    );
-                }
-            }
-        }
-
-        if needs_publish {
-            match self
-                .create_and_publish_key_package(account, key_package_relays)
-                .await
-            {
-                Ok(()) => {
-                    tracing::debug!(target: "whitenoise::accounts", "Published key package");
-                }
-                Err(e) => {
-                    tracing::warn!(
-                        target: "whitenoise::accounts",
-                        "Initial key package publish failed, scheduler will retry: {}",
-                        e
-                    );
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Checks whether a key package event was published by this Whitenoise instance
-    /// and still has live local key material.
-    async fn is_own_key_package(&self, pubkey: &PublicKey, event_id: &EventId) -> bool {
-        match PublishedKeyPackage::find_by_event_id(pubkey, &event_id.to_hex(), &self.database)
-            .await
-        {
-            Ok(Some(pkg)) => !pkg.key_material_deleted,
-            Ok(None) => false,
-            Err(e) => {
-                tracing::warn!(
-                    target: "whitenoise::accounts",
-                    "Failed to look up published key package for event {}: {}",
-                    event_id,
-                    e
-                );
-                false
-            }
-        }
-    }
-
-    async fn load_default_relays(&self) -> Result<Vec<Relay>> {
-        let mut default_relays = Vec::new();
-        for Relay { url, .. } in Relay::defaults() {
-            let relay = self.find_or_create_relay_by_url(&url).await?;
-            default_relays.push(relay);
-        }
-        Ok(default_relays)
-    }
-
-    /// Sets up the relays for a new account.
-    ///
-    /// # Arguments
-    ///
-    /// * `account` - The account to setup the relays for
-    /// * `user` - The user to setup the relays for
-    ///
-    /// # Returns
-    /// Returns the default relays for the account.
-    async fn setup_relays_for_new_account(
-        &self,
-        account: &mut Account,
-        user: &User,
-    ) -> Result<Vec<Relay>> {
-        let default_relays = self.load_default_relays().await?;
-
-        user.add_relays(&default_relays, RelayType::Nip65, &self.database)
-            .await?;
-        user.add_relays(&default_relays, RelayType::Inbox, &self.database)
-            .await?;
-        user.add_relays(&default_relays, RelayType::KeyPackage, &self.database)
-            .await?;
-
-        self.background_publish_account_relay_list(
-            account,
-            RelayType::Nip65,
-            Some(&default_relays),
-        )
-        .await?;
-        self.background_publish_account_relay_list(
-            account,
-            RelayType::Inbox,
-            Some(&default_relays),
-        )
-        .await?;
-        self.background_publish_account_relay_list(
-            account,
-            RelayType::KeyPackage,
-            Some(&default_relays),
-        )
-        .await?;
-
-        Ok(default_relays)
-    }
-
-    async fn setup_relays_for_existing_account(
-        &self,
-        account: &mut Account,
-    ) -> Result<(Vec<Relay>, Vec<Relay>, Vec<Relay>)> {
-        let default_relays = self.load_default_relays().await?;
-        let signer = self.get_signer_for_account(account)?;
-
-        // NIP-65 must be fetched first because Inbox and KeyPackage use the
-        // discovered NIP-65 relays as their query source.
-        let (nip65_relays, should_publish_nip65) = self
-            .setup_existing_account_relay_type(
-                account,
-                RelayType::Nip65,
-                &default_relays,
-                &default_relays,
-            )
-            .await?;
-
-        // Inbox and KeyPackage fetches are independent — run them concurrently.
-        let (inbox_result, key_package_result) = tokio::join!(
-            self.setup_existing_account_relay_type(
-                account,
-                RelayType::Inbox,
-                &nip65_relays,
-                &default_relays,
-            ),
-            self.setup_existing_account_relay_type(
-                account,
-                RelayType::KeyPackage,
-                &nip65_relays,
-                &default_relays,
-            ),
-        );
-        let (inbox_relays, should_publish_inbox) = inbox_result?;
-        let (key_package_relays, should_publish_key_package) = key_package_result?;
-
-        // Publish any relay lists that need publishing concurrently.
-        let nip65_publish = async {
-            if should_publish_nip65 {
-                self.publish_relay_list(
-                    &nip65_relays,
-                    RelayType::Nip65,
-                    &nip65_relays,
-                    signer.clone(),
-                )
-                .await
-            } else {
-                Ok(())
-            }
-        };
-        let inbox_publish = async {
-            if should_publish_inbox {
-                self.publish_relay_list(
-                    &inbox_relays,
-                    RelayType::Inbox,
-                    &nip65_relays,
-                    signer.clone(),
-                )
-                .await
-            } else {
-                Ok(())
-            }
-        };
-        let key_package_publish = async {
-            if should_publish_key_package {
-                self.publish_relay_list(
-                    &key_package_relays,
-                    RelayType::KeyPackage,
-                    &nip65_relays,
-                    signer.clone(),
-                )
-                .await
-            } else {
-                Ok(())
-            }
-        };
-        let (r1, r2, r3) = tokio::join!(nip65_publish, inbox_publish, key_package_publish);
-        r1?;
-        r2?;
-        r3?;
-
-        Ok((nip65_relays, inbox_relays, key_package_relays))
-    }
-
-    /// Sets up relays for an external signer account.
-    ///
-    /// This is similar to `setup_relays_for_existing_account` but does NOT publish
-    /// relay lists (the caller with the signer should handle that).
-    /// It only fetches existing relays or uses defaults and saves them locally.
-    ///
-    /// Returns the relays for each type and booleans indicating which relay lists
-    /// should be published (true when defaults were used).
-    async fn setup_relays_for_external_signer_account(
-        &self,
-        account: &mut Account,
-    ) -> Result<ExternalSignerRelaySetup> {
-        let default_relays = self.load_default_relays().await?;
-
-        // NIP-65 must be fetched first because Inbox and KeyPackage use the
-        // discovered NIP-65 relays as their query source.
-        let (nip65_relays, should_publish_nip65) = self
-            .setup_external_account_relay_type(
-                account,
-                RelayType::Nip65,
-                &default_relays,
-                &default_relays,
-            )
-            .await?;
-
-        // Inbox and KeyPackage fetches are independent — run them concurrently.
-        let (inbox_result, key_package_result) = tokio::join!(
-            self.setup_external_account_relay_type(
-                account,
-                RelayType::Inbox,
-                &nip65_relays,
-                &default_relays,
-            ),
-            self.setup_external_account_relay_type(
-                account,
-                RelayType::KeyPackage,
-                &nip65_relays,
-                &default_relays,
-            ),
-        );
-        let (inbox_relays, should_publish_inbox) = inbox_result?;
-        let (key_package_relays, should_publish_key_package) = key_package_result?;
-
-        Ok(ExternalSignerRelaySetup {
-            nip65_relays,
-            inbox_relays,
-            key_package_relays,
-            should_publish_nip65,
-            should_publish_inbox,
-            should_publish_key_package,
-        })
-    }
-
-    /// Sets up a specific relay type for an external signer account.
-    /// Fetches from network or uses defaults, but never publishes.
-    ///
-    /// Returns the relays and a boolean indicating whether they should be published
-    /// (true if defaults were used because no existing relays were found).
-    async fn setup_external_account_relay_type(
-        &self,
-        account: &Account,
-        relay_type: RelayType,
-        source_relays: &[Relay],
-        default_relays: &[Relay],
-    ) -> Result<(Vec<Relay>, bool)> {
-        // Try to fetch existing relay lists first
-        let fetched_relays = self
-            .fetch_existing_relays(account.pubkey, relay_type, source_relays)
-            .await?;
-
-        let user = account.user(&self.database).await?;
-        if fetched_relays.is_empty() {
-            // No existing relay lists - use defaults, mark for publishing
-            user.add_relays(default_relays, relay_type, &self.database)
-                .await?;
-            Ok((default_relays.to_vec(), true))
-        } else {
-            // Found existing relay lists - use them, no publishing needed
-            user.add_relays(&fetched_relays, relay_type, &self.database)
-                .await?;
-            Ok((fetched_relays, false))
-        }
-    }
-
-    async fn setup_existing_account_relay_type(
-        &self,
-        account: &Account,
-        relay_type: RelayType,
-        source_relays: &[Relay],
-        default_relays: &[Relay],
-    ) -> Result<(Vec<Relay>, bool)> {
-        // Existing accounts: try to fetch existing relay lists first
-        let fetched_relays = self
-            .fetch_existing_relays(account.pubkey, relay_type, source_relays)
-            .await?;
-
-        if fetched_relays.is_empty() {
-            // No existing relay lists - use defaults and publish
-            self.add_relays_to_account(account, default_relays, relay_type)
-                .await?;
-            Ok((default_relays.to_vec(), true))
-        } else {
-            // Found existing relay lists - use them, no publishing needed
-            let user = account.user(&self.database).await?;
-            user.add_relays(&fetched_relays, relay_type, &self.database)
-                .await?;
-            Ok((fetched_relays, false))
-        }
-    }
-
-    async fn fetch_existing_relays(
-        &self,
-        pubkey: PublicKey,
-        relay_type: RelayType,
-        source_relays: &[Relay],
-    ) -> Result<Vec<Relay>> {
-        let source_relay_urls = Relay::urls(source_relays);
-        // Ensure source relays are in the client pool before attempting to fetch.
-        // Without this, fetch_events_from will fail with RelayNotFound if the
-        // relays haven't been added to the pool yet (e.g. during login before
-        // activate_account runs).
-        self.nostr
-            .ensure_relays_connected(&source_relay_urls)
-            .await?;
-        let relay_event = self
-            .nostr
-            .fetch_user_relays(pubkey, relay_type, &source_relay_urls)
-            .await?;
-
-        let mut relays = Vec::new();
-        if let Some(event) = relay_event {
-            let relay_urls = NostrManager::relay_urls_from_event(&event);
-            for url in relay_urls {
-                let relay = self.find_or_create_relay_by_url(&url).await?;
-                relays.push(relay);
-            }
-        }
-
-        Ok(relays)
-    }
-
-    async fn add_relays_to_account(
-        &self,
-        account: &Account,
-        relays: &[Relay],
-        relay_type: RelayType,
-    ) -> Result<()> {
-        if relays.is_empty() {
-            return Ok(());
-        }
-
-        let user = account.user(&self.database).await?;
-        user.add_relays(relays, relay_type, &self.database).await?;
-
-        self.background_publish_account_relay_list(account, relay_type, Some(relays))
-            .await?;
-
-        tracing::debug!(target: "whitenoise::accounts", "Added {} relays of type {:?} to account", relays.len(), relay_type);
-
-        Ok(())
-    }
-
-    async fn publish_relay_list<S>(
-        &self,
-        relays: &[Relay],
-        relay_type: RelayType,
-        target_relays: &[Relay],
-        signer: S,
-    ) -> Result<()>
-    where
-        S: NostrSigner + 'static,
-    {
-        let relays_urls = Relay::urls(relays);
-        let target_relays_urls = Relay::urls(target_relays);
-        self.nostr
-            .publish_relay_list_with_signer(&relays_urls, relay_type, &target_relays_urls, signer)
-            .await?;
-        Ok(())
-    }
-
-    pub(crate) async fn background_publish_account_metadata(
-        &self,
-        account: &Account,
-    ) -> Result<()> {
-        let account_clone = account.clone();
-        let nostr = self.nostr.clone();
-        let signer = self.get_signer_for_account(account)?;
-        let user = account.user(&self.database).await?;
-        let relays = account.nip65_relays(self).await?;
-
-        tokio::spawn(async move {
-            tracing::debug!(target: "whitenoise::accounts", "Background task: Publishing metadata for account: {:?}", account_clone.pubkey);
-
-            let relays_urls = Relay::urls(&relays);
-
-            nostr
-                .publish_metadata_with_signer(&user.metadata, &relays_urls, signer)
-                .await?;
-
-            tracing::debug!(target: "whitenoise::accounts", "Successfully published metadata for account: {:?}", account_clone.pubkey);
-            Ok::<(), WhitenoiseError>(())
-        });
-        Ok(())
-    }
-
-    /// Publishes the relay list for the account to the Nostr network.
-    ///
-    /// # Arguments
-    ///
-    /// * `account` - The account to publish the relay list for
-    /// * `relay_type` - The type of relay list to publish
-    /// * `relays` - The relays to publish the relay list to, if None, the relays will be fetched from the account
-    pub(crate) async fn background_publish_account_relay_list(
-        &self,
-        account: &Account,
-        relay_type: RelayType,
-        relays: Option<&[Relay]>,
-    ) -> Result<()> {
-        let account_clone = account.clone();
-        let nostr = self.nostr.clone();
-        let relays = if let Some(relays) = relays {
-            relays.to_vec()
-        } else {
-            account.relays(relay_type, self).await?
-        };
-        let signer = self.get_signer_for_account(account)?;
-        let target_relays = if relay_type == RelayType::Nip65 {
-            relays.clone()
-        } else {
-            account.nip65_relays(self).await?
-        };
-
-        tokio::spawn(async move {
-            tracing::debug!(target: "whitenoise::accounts", "Background task: Publishing relay list for account: {:?}", account_clone.pubkey);
-
-            let relays_urls = Relay::urls(&relays);
-            let target_relays_urls = Relay::urls(&target_relays);
-
-            nostr
-                .publish_relay_list_with_signer(
-                    &relays_urls,
-                    relay_type,
-                    &target_relays_urls,
-                    signer,
-                )
-                .await?;
-
-            tracing::debug!(target: "whitenoise::accounts", "Successfully published relay list for account: {:?}", account_clone.pubkey);
-            Ok::<(), WhitenoiseError>(())
-        });
-        Ok(())
-    }
-
-    pub(crate) async fn background_publish_account_follow_list(
-        &self,
-        account: &Account,
-    ) -> Result<()> {
-        let account_clone = account.clone();
-        let nostr = self.nostr.clone();
-        let relays = account.nip65_relays(self).await?;
-        let signer = self.get_signer_for_account(account)?;
-        let follows = account.follows(&self.database).await?;
-        let follows_pubkeys = follows.iter().map(|f| f.pubkey).collect::<Vec<_>>();
-
-        tokio::spawn(async move {
-            tracing::debug!(target: "whitenoise::accounts", "Background task: Publishing follow list for account: {:?}", account_clone.pubkey);
-
-            let relays_urls = Relay::urls(&relays);
-            nostr
-                .publish_follow_list_with_signer(&follows_pubkeys, &relays_urls, signer)
-                .await?;
-
-            tracing::debug!(target: "whitenoise::accounts", "Successfully published follow list for account: {:?}", account_clone.pubkey);
-            Ok::<(), WhitenoiseError>(())
-        });
-        Ok(())
-    }
-
-    /// Extract group data including relay URLs and group IDs for subscription setup.
-    pub(crate) async fn extract_groups_relays_and_ids(
-        &self,
-        account: &Account,
-    ) -> Result<(Vec<RelayUrl>, Vec<String>)> {
-        let mdk = self.create_mdk_for_account(account.pubkey)?;
-        let groups = mdk.get_groups()?;
-        let mut group_relays_set = HashSet::new();
-        let mut group_ids = vec![];
-
-        for group in &groups {
-            let relays = mdk.get_relays(&group.mls_group_id)?;
-            group_relays_set.extend(relays);
-            group_ids.push(hex::encode(group.nostr_group_id));
-        }
-
-        let group_relays_urls = group_relays_set.into_iter().collect::<Vec<_>>();
-
-        Ok((group_relays_urls, group_ids))
-    }
-
-    pub(crate) async fn setup_subscriptions(
-        &self,
-        account: &Account,
-        nip65_relays: &[Relay],
-        inbox_relays: &[Relay],
-    ) -> Result<()> {
-        tracing::debug!(
-            target: "whitenoise::accounts",
-            "Setting up subscriptions for account: {:?}",
-            account
-        );
-
-        let user_relays: Vec<RelayUrl> = Relay::urls(nip65_relays);
-
-        let inbox_relays: Vec<RelayUrl> = Relay::urls(inbox_relays);
-
-        let (group_relays_urls, nostr_group_ids) =
-            self.extract_groups_relays_and_ids(account).await?;
-
-        // Ensure group relays are in the database
-        for relay_url in &group_relays_urls {
-            Relay::find_or_create_by_url(relay_url, &self.database).await?;
-        }
-
-        // Compute per-account since with a 10s lookback buffer when available
-        let since = account.since_timestamp(10);
-        match since {
-            Some(ts) => tracing::debug!(
-                target: "whitenoise::accounts",
-                "Computed per-account since={}s (10s buffer) for {}",
-                ts.as_secs(),
-                account.pubkey.to_hex()
-            ),
-            None => tracing::debug!(
-                target: "whitenoise::accounts",
-                "Computed per-account since=None (unsynced) for {}",
-                account.pubkey.to_hex()
-            ),
-        }
-
-        let signer = self.get_signer_for_account(account)?;
-        self.nostr
-            .setup_account_subscriptions_with_signer(
-                account.pubkey,
-                &user_relays,
-                &inbox_relays,
-                &group_relays_urls,
-                &nostr_group_ids,
-                since,
-                signer,
-            )
-            .await?;
-
-        tracing::debug!(
-            target: "whitenoise::accounts",
-            "Subscriptions setup"
-        );
-        Ok(())
-    }
-
-    /// Refresh account subscriptions.
-    ///
-    /// This method updates subscriptions when account state changes (group membership, relay preferences).
-    /// Uses explicit cleanup to handle relay changes properly - NIP-01 auto-replacement only works
-    /// within the same relay, so changing relays would leave orphaned subscriptions without cleanup.
-    ///
-    /// # Arguments
-    ///
-    /// * `account` - The account to refresh subscriptions for
-    pub(crate) async fn refresh_account_subscriptions(&self, account: &Account) -> Result<()> {
-        tracing::debug!(
-            target: "whitenoise::accounts",
-            "Refreshing account subscriptions for account: {:?}",
-            account.pubkey
-        );
-
-        let user_relays: Vec<RelayUrl> = Relay::urls(&account.nip65_relays(self).await?);
-
-        let inbox_relays: Vec<RelayUrl> = Relay::urls(&account.inbox_relays(self).await?);
-
-        let (group_relays_urls, nostr_group_ids) =
-            self.extract_groups_relays_and_ids(account).await?;
-
-        let signer = self.get_signer_for_account(account)?;
-        self.nostr
-            .update_account_subscriptions_with_signer(
-                account.pubkey,
-                &user_relays,
-                &inbox_relays,
-                &group_relays_urls,
-                &nostr_group_ids,
-                signer,
-            )
-            .await
-            .map_err(WhitenoiseError::from)
-    }
-}
-
-#[cfg(test)]
-pub mod test_utils {
-    use mdk_core::MDK;
-    use mdk_sqlite_storage::MdkSqliteStorage;
-    use nostr_sdk::PublicKey;
-    use std::path::PathBuf;
-    use tempfile::TempDir;
-
-    pub fn data_dir() -> PathBuf {
-        TempDir::new().unwrap().path().to_path_buf()
-    }
-
-    pub fn create_mdk(pubkey: PublicKey) -> MDK<MdkSqliteStorage> {
-        super::super::Whitenoise::initialize_mock_keyring_store();
-        super::Account::create_mdk(pubkey, &data_dir(), "com.whitenoise.test").unwrap()
-    }
 }
 
 #[cfg(test)]
@@ -2905,82 +1696,6 @@ mod tests {
             assert_eq!(published_metadata.display_name, new_metadata.display_name);
             assert_eq!(published_metadata.about, new_metadata.about);
         }
-    }
-
-    #[tokio::test]
-    async fn test_extract_groups_relays_and_ids_no_groups() {
-        let (whitenoise, _data_temp, _logs_temp) = create_mock_whitenoise().await;
-        let account = whitenoise.create_identity().await.unwrap();
-
-        let (relays, group_ids) = whitenoise
-            .extract_groups_relays_and_ids(&account)
-            .await
-            .unwrap();
-
-        assert!(
-            relays.is_empty(),
-            "Should have no relays when account has no groups"
-        );
-        assert!(
-            group_ids.is_empty(),
-            "Should have no group IDs when account has no groups"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_extract_groups_relays_and_ids_with_groups() {
-        let (whitenoise, _data_temp, _logs_temp) = create_mock_whitenoise().await;
-
-        // Create creator and member accounts
-        let creator_account = whitenoise.create_identity().await.unwrap();
-        let member_account = whitenoise.create_identity().await.unwrap();
-
-        // Allow time for key packages to be published
-        tokio::time::sleep(tokio::time::Duration::from_millis(200)).await;
-
-        let relay1 = RelayUrl::parse("ws://localhost:8080").unwrap();
-        let relay2 = RelayUrl::parse("ws://localhost:7777").unwrap();
-
-        // Create a group with specific relays
-        let config = NostrGroupConfigData::new(
-            "Test Group".to_string(),
-            "Test Description".to_string(),
-            None,
-            None,
-            None,
-            vec![relay1.clone(), relay2.clone()],
-            vec![creator_account.pubkey],
-        );
-
-        let group = whitenoise
-            .create_group(&creator_account, vec![member_account.pubkey], config, None)
-            .await
-            .unwrap();
-
-        // Extract groups relays and IDs
-        let (relays, group_ids) = whitenoise
-            .extract_groups_relays_and_ids(&creator_account)
-            .await
-            .unwrap();
-
-        // Verify relays were extracted
-        assert!(!relays.is_empty(), "Should have relays from the group");
-        assert!(
-            relays.contains(&relay1),
-            "Should contain relay1 from group config"
-        );
-        assert!(
-            relays.contains(&relay2),
-            "Should contain relay2 from group config"
-        );
-
-        // Verify group ID was extracted
-        assert_eq!(group_ids.len(), 1, "Should have one group ID");
-        assert_eq!(
-            group_ids[0],
-            hex::encode(group.nostr_group_id),
-            "Group ID should match the created group"
-        );
     }
 
     #[tokio::test]
@@ -4995,80 +3710,6 @@ mod tests {
         assert_eq!(result.status, LoginStatus::Complete);
         assert_eq!(result.account.pubkey, pubkey);
         assert_eq!(result.account.account_type, AccountType::External);
-    }
-
-    #[tokio::test]
-    async fn test_is_own_key_package_returns_false_for_unknown_event() {
-        let (whitenoise, _data_temp, _logs_temp) = create_mock_whitenoise().await;
-        let (account, _keys) = create_test_account(&whitenoise).await;
-
-        let event_id = EventId::all_zeros();
-        assert!(
-            !whitenoise
-                .is_own_key_package(&account.pubkey, &event_id)
-                .await
-        );
-    }
-
-    #[tokio::test]
-    async fn test_is_own_key_package_returns_true_for_tracked_kp() {
-        let (whitenoise, _data_temp, _logs_temp) = create_mock_whitenoise().await;
-        let (account, _keys) = create_test_account(&whitenoise).await;
-        account.save(&whitenoise.database).await.unwrap();
-
-        // Use a valid 64-char hex string as event_id
-        let event_hex = "abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890";
-        PublishedKeyPackage::create(&account.pubkey, &[1, 2, 3], event_hex, &whitenoise.database)
-            .await
-            .unwrap();
-
-        let event_id = EventId::parse(event_hex).unwrap();
-        assert!(
-            whitenoise
-                .is_own_key_package(&account.pubkey, &event_id)
-                .await
-        );
-    }
-
-    #[tokio::test]
-    async fn test_is_own_key_package_returns_false_when_key_material_deleted() {
-        let (whitenoise, _data_temp, _logs_temp) = create_mock_whitenoise().await;
-        let (account, _keys) = create_test_account(&whitenoise).await;
-        account.save(&whitenoise.database).await.unwrap();
-
-        let event_hex = "abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890";
-        PublishedKeyPackage::create(&account.pubkey, &[1, 2, 3], event_hex, &whitenoise.database)
-            .await
-            .unwrap();
-
-        // Mark key material as deleted
-        let pkg =
-            PublishedKeyPackage::find_by_event_id(&account.pubkey, event_hex, &whitenoise.database)
-                .await
-                .unwrap()
-                .unwrap();
-        PublishedKeyPackage::mark_key_material_deleted(pkg.id, &whitenoise.database)
-            .await
-            .unwrap();
-
-        let event_id = EventId::parse(event_hex).unwrap();
-        assert!(
-            !whitenoise
-                .is_own_key_package(&account.pubkey, &event_id)
-                .await
-        );
-    }
-
-    #[tokio::test]
-    async fn test_is_own_key_package_returns_false_on_db_error() {
-        let (whitenoise, _data_temp, _logs_temp) = create_mock_whitenoise().await;
-        let pubkey = Keys::generate().public_key();
-
-        // Close the pool to simulate a database error
-        whitenoise.database.pool.close().await;
-
-        let event_id = EventId::all_zeros();
-        assert!(!whitenoise.is_own_key_package(&pubkey, &event_id).await);
     }
 
     // -----------------------------------------------------------------------
